Split KtorClient to separate module
