+++ jellyfin-core/build.gradle.kts
@@ -21,6 +21,7 @@ kotlin {
 		val commonMain by getting {
 			dependencies {
 				api(projects.jellyfinApi)
+				api(projects.jellyfinApiKtor)
 				api(projects.jellyfinModel)
 
 				implementation(libs.kotlinx.coroutines)
