+++ jellyfin-core/src/jvmMain/kotlin/org/jellyfin/sdk/JellyfinOptions.kt
@@ -1,6 +1,6 @@
 package org.jellyfin.sdk
 
-import org.jellyfin.sdk.api.client.KtorClient
+import org.jellyfin.sdk.api.ktor.KtorClient
 import org.jellyfin.sdk.api.sockets.OkHttpWebsocketSession
 import org.jellyfin.sdk.api.sockets.SocketConnectionFactory
 import org.jellyfin.sdk.model.ClientInfo
